requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
bitsandbytes==0.45.5
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.13.2
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.82.0
dataclasses-json==0.6.7
datasets==4.5.0
definitions==0.2.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
dynaconf==3.2.12
evidently==0.7.20
Faker==40.1.2
fastapi==0.128.0
fastuuid==0.14.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huey==2.6.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.0
langchain-anthropic==1.3.1
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.7
langchain-groq==1.1.0
langchain-neo4j==0.6.0
langchain-openai==1.1.7
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.46
litestar==2.19.0
litestar-htmx==0.5.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mmh3==5.2.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multipart==1.3.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.13.0
neo4j==5.28.2
neo4j-graphrag==1.10.1
networkx==3.6
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.15.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
peft==0.18.1
pillow==12.0.0
platformdirs==4.9.1
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.2.0
posthog==5.4.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==3.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-click==1.9.6
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
smolagents==1.24.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlite-vec==0.1.6
sqlparse==0.5.5
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.52.0
sympy==1.14.0
tavily-python==0.7.21
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.2
tqdm==4.67.1
transformers==5.1.0
trl==0.28.0
typer==0.23.1
typer-slim==0.23.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.3.0
uuid6==2025.0.1
uuid_utils==0.13.0
uvicorn==0.38.0
waitress==3.0.2
wandb==0.25.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
xgboost==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
